+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 
 ## [Unreleased]
+- Add `<foo {..iter}>` syntax for dynamic attributes
 
 ## 0.1.0 - 2023-11-14
 ### Added